Fixed code error in `resource_pack_manager.cpp`. Added hook

+++ resedit/core/common.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 #define RESEDIT_RESOURCE_PACK_METADATA_FILENAME "metadata.json"
 #define RESEDIT_RESOURCE_PACK_DEFAULT_NAME "Unknown pack"
@@ -11,6 +12,6 @@ struct AssetData
 {
 	std::string const path;
 	char* const buffer;
-	unsigned const max_len;
-	unsigned* const written_len;
+	uint64_t const max_len;
+	uint64_t* const written_len;
 };
